[ZH] Fixed warnings about signed/unsigned integer comparison (#503)

+++ GeneralsMD/Code/GameEngineDevice/Source/W3DDevice/GameClient/Drawable/Draw/W3DModelDraw.cpp
@@ -3332,7 +3332,7 @@ Bool W3DModelDraw::getProjectileLaunchOffset(
 	}
 	else
 	{
-		if (specificBarrelToUse < 0 || specificBarrelToUse >= wbvec.size())
+		if (specificBarrelToUse < 0 || specificBarrelToUse >= static_cast<long>(wbvec.size()))
 			specificBarrelToUse = 0;
 
 		if (launchPos)
